Refactoring & merge

+++ core/src/main/java/dev/muskrat/delivery/dto/product/ProductDTO.java
@@ -14,5 +14,5 @@ public class ProductDTO {
     private String url;
     private Boolean available;
     private Double value;
-    private Integer category;
+    private Long category;
 }
